Update: Several logic fixes and enhancements
Instead of using EID to auto-open an exercise program, we will now use the exercise code. The UID that the custom supplies can now be a string or integer. Updated the API to support this as well. The register function was getting called after rendering functions. Updated the logic to register the user, and if registration is successful, then continue with the plugin init.