Bug Fixes: Corrected several filter system bugs
Implemented the Custom Favorites category filter system. The condition drop down is now displayed as the sub-category drop down. The other filters are hidden and not used for the custom categories (except for the equipment drop down). Also added additional callback functions for tracking. The video_bg option can now be set and will affect the exercise videos on the viewer screen. Also made the switch from exercise program ID to CODE for the direct program link.