Bug Fixes
DV-6736 - Fixed an issue with clearing the mobile search field and reloading the result set correctly. Also fixed the missing clear icon in the mobile search field.
DV-6728 - Fixed Anatomical diagram overlaying exercise details. Added a min-height in conjunction with height 100% to ensure the containing div for the anatomical model is tall enough.